+++ src/module_system/import_resolution/stdlib_gates.rs
@@ -1,0 +1,73 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum GatedStdlibModule {
+    ActorFramework,
+    AllocatorFramework,
+    AsyncRuntime,
+    SyncRuntime,
+}
+
+impl GatedStdlibModule {
+    const CONCURRENCY_SEGMENT: &'static str = "concurrency";
+    const ACTOR_SEGMENT: &'static str = "actor";
+    const ASYNC_SEGMENT: &'static str = "async";
+    const SYNC_SEGMENT: &'static str = "sync";
+    const MEMORY_SEGMENT: &'static str = "memory";
+    const ALLOCATOR_SEGMENT: &'static str = "allocator";
+
+    pub(super) fn from_sub_path(sub_path: &[String]) -> Option<Self> {
+        if sub_path
+            .first()
+            .is_some_and(|segment| segment == Self::CONCURRENCY_SEGMENT)
+            && sub_path
+                .get(1)
+                .is_some_and(|segment| segment == Self::ACTOR_SEGMENT)
+        {
+            return Some(Self::ActorFramework);
+        }
+        if sub_path
+            .first()
+            .is_some_and(|segment| segment == Self::CONCURRENCY_SEGMENT)
+            && sub_path
+                .get(1)
+                .is_some_and(|segment| segment == Self::ASYNC_SEGMENT)
+        {
+            return Some(Self::AsyncRuntime);
+        }
+        if sub_path
+            .first()
+            .is_some_and(|segment| segment == Self::CONCURRENCY_SEGMENT)
+            && sub_path
+                .get(1)
+                .is_some_and(|segment| segment == Self::SYNC_SEGMENT)
+        {
+            return Some(Self::SyncRuntime);
+        }
+        if sub_path
+            .first()
+            .is_some_and(|segment| segment == Self::MEMORY_SEGMENT)
+            && sub_path
+                .get(1)
+                .is_some_and(|segment| segment == Self::ALLOCATOR_SEGMENT)
+        {
+            return Some(Self::AllocatorFramework);
+        }
+        None
+    }
+
+    pub(super) fn gate_message(self) -> &'static str {
+        match self {
+            Self::ActorFramework => {
+                "std actor framework modules are gated until mailbox, scheduling, supervisor, and allocator semantics are implemented"
+            }
+            Self::AllocatorFramework => {
+                "std allocator modules are gated until allocator ownership and effect semantics are implemented"
+            }
+            Self::AsyncRuntime => {
+                "std async runtime modules are gated until Sync/Async effect checking and task lowering are implemented"
+            }
+            Self::SyncRuntime => {
+                "std sync runtime modules are gated until channel, mailbox, and blocking semantics are implemented"
+            }
+        }
+    }
+}
